replace the balance field from f64 to decimal to prevent the precision problem

+++ src/models/account.rs
@@ -1,4 +1,5 @@
 use chrono::{DateTime, Utc};
+use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 use sqlx::FromRow;
 use uuid::Uuid;
@@ -7,7 +8,7 @@ use uuid::Uuid;
 pub struct Account {
     pub id: Uuid,
     pub user_id: Uuid,
-    pub balance: f64,
+    pub balance: Decimal,
     pub created_at: DateTime<Utc>,
     pub updated_at: DateTime<Utc>,
 }
